restructure logic
